+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+### Updated
+- Update `ExistsEntity()` to receive `table` as parameter
+
 ## [1.0.0]
 ### Added
 - Thin layers around SlicingDice API endpoints